+++ temba/contacts/search/omnibox.py
@@ -7,7 +7,6 @@ from django.db.models.functions import Upper
 
 from temba.channels.models import Channel
 from temba.contacts.models import Contact, ContactGroup, ContactGroupCount, ContactURN
-from temba.msgs.models import Label
 from temba.utils.models.es import IDSliceQuerySet
 
 from . import SearchException, search_contacts
@@ -24,37 +23,25 @@ def omnibox_query(org, **kwargs):
     """
     # determine what type of group/contact/URN lookup is being requested
     contact_uuids = kwargs.get("c", None)  # contacts with ids
-    message_ids = kwargs.get("m", None)  # contacts with message ids
-    label_id = kwargs.get("l", None)  # contacts in flow step with UUID
     group_uuids = kwargs.get("g", None)  # groups with ids
     urn_ids = kwargs.get("u", None)  # URNs with ids
     search = kwargs.get("search", None)  # search of groups, contacts and URNs
     types = list(kwargs.get("types", ""))  # limit search to types (g | s | c | u)
 
-    # these lookups return a Contact queryset
-    if contact_uuids or message_ids or label_id:
-        qs = Contact.objects.filter(org=org, status=Contact.STATUS_ACTIVE, is_active=True)
+    if contact_uuids:
+        return (
+            Contact.objects.filter(
+                org=org, status=Contact.STATUS_ACTIVE, is_active=True, uuid__in=contact_uuids.split(",")
+            )
+            .distinct()
+            .order_by("name")
+        )
 
-        if contact_uuids:
-            qs = qs.filter(uuid__in=contact_uuids.split(","))
-
-        elif message_ids:
-            qs = qs.filter(msgs__in=message_ids.split(","))
-
-        elif label_id:
-            label = Label.label_objects.get(pk=label_id)
-            qs = qs.filter(msgs__in=label.get_messages())
-
-        return qs.distinct().order_by("name")
-
-    # this lookup returns a ContactGroup queryset
     elif group_uuids:
         return ContactGroup.get_groups(org).filter(uuid__in=group_uuids.split(",")).order_by("name")
 
-    # this lookup returns a ContactURN queryset
     elif urn_ids:
-        qs = ContactURN.objects.filter(org=org, id__in=urn_ids.split(",")).select_related("contact")
-        return qs.order_by("path")
+        return ContactURN.objects.filter(org=org, id__in=urn_ids.split(",")).select_related("contact").order_by("path")
 
     # searching returns something which acts enough like a queryset to be paged
     return omnibox_mixed_search(org, search, types)