+++ temba/tickets/models.py
@@ -1,4 +1,7 @@
 from abc import ABCMeta
+from datetime import date
+
+import openpyxl
 
 from django.conf import settings
 from django.contrib.auth.models import User
@@ -13,6 +16,7 @@ from django.utils.translation import gettext_lazy as _
 from temba import mailroom
 from temba.contacts.models import Contact
 from temba.orgs.models import DependencyMixin, Org, UserSettings
+from temba.utils.dates import date_range
 from temba.utils.models import DailyCountModel, SquashableModel, TembaModel
 from temba.utils.uuid import uuid4
 
@@ -507,6 +511,60 @@ class TicketDailyCount(DailyCountModel):
     @classmethod
     def get_by_users(cls, org, users, count_type: str, since=None, until=None):
         return cls._get_count_set(count_type, {f"o:{org.id}:u:{u.id}": u for u in users}, since, until)
+
+    @classmethod
+    def export_summary(cls, org: Org, since: date, until: date) -> openpyxl.Workbook:
+        workbook = openpyxl.Workbook()
+        sheet = workbook.active
+        sheet.title = "Tickets"
+        sheet.merge_cells("A1:A2")
+        sheet.cell(row=1, column=2, value="Workspace")
+        sheet.merge_cells("B1:C1")
+        sheet.cell(row=2, column=2, value="Opened")
+        sheet.cell(row=2, column=3, value="Replies")
+
+        users = list(org.get_users().order_by("email"))
+
+        user_col = 4
+        for user in users:
+            cell = sheet.cell(row=1, column=user_col, value=str(user))
+            cell.hyperlink = f"mailto:{user.email}"
+            cell.style = "Hyperlink"
+            sheet.merge_cells(start_row=1, start_column=user_col, end_row=1, end_column=user_col + 1)
+
+            sheet.cell(row=2, column=user_col, value="Assigned")
+            sheet.cell(row=2, column=user_col + 1, value="Replies")
+            user_col += 2
+
+        def by_day(cs: list) -> dict:
+            return {c[0]: c[1] for c in cs}
+
+        org_openings = by_day(cls.get_by_org(org, cls.TYPE_OPENING, since, until).day_totals())
+        org_replies = by_day(cls.get_by_org(org, cls.TYPE_REPLY, since, until).day_totals())
+
+        user_assignments = {}
+        user_replies = {}
+        for user in users:
+            user_assignments[user] = by_day(
+                cls.get_by_users(org, [user], cls.TYPE_ASSIGNMENT, since, until).day_totals()
+            )
+            user_replies[user] = by_day(cls.get_by_users(org, [user], cls.TYPE_REPLY, since, until).day_totals())
+
+        day_row = 3
+        for day in date_range(since, until):
+            sheet.cell(row=day_row, column=1, value=day)
+            sheet.cell(row=day_row, column=2, value=org_openings.get(day, 0))
+            sheet.cell(row=day_row, column=3, value=org_replies.get(day, 0))
+
+            user_col = 4
+            for user in users:
+                sheet.cell(row=day_row, column=user_col, value=user_assignments[user].get(day, 0))
+                sheet.cell(row=day_row, column=user_col + 1, value=user_replies[user].get(day, 0))
+                user_col += 2
+
+            day_row += 1
+
+        return workbook
 
     class Meta:
         indexes = [