+++ temba/utils/tests.py
@@ -3,6 +3,7 @@ import datetime
 import io
 import os
 from collections import OrderedDict
+from datetime import date
 from decimal import Decimal
 from unittest.mock import PropertyMock, patch
 
@@ -29,7 +30,7 @@ from temba.utils.templatetags.temba import format_datetime, icon
 from . import chunk_list, countries, format_number, languages, percentage, redact, sizeof_fmt, str_to_bool
 from .cache import get_cacheable_attr, get_cacheable_result, incrby_existing
 from .celery import nonoverlapping_task
-from .dates import datetime_to_str, datetime_to_timestamp, timestamp_to_datetime
+from .dates import date_range, datetime_to_str, datetime_to_timestamp, timestamp_to_datetime
 from .email import is_valid_address, send_simple_email
 from .export import TableExporter
 from .fields import NameValidator, validate_external_url
@@ -234,6 +235,13 @@ class DatesTest(TembaTest):
         self.assertIsNone(datetime_to_str(None, "%Y-%m-%d %H:%M", tz=tz))
         self.assertEqual(datetime_to_str(d2, "%Y-%m-%d %H:%M", tz=tz), "2014-01-02 03:04")
         self.assertEqual(datetime_to_str(d2, "%Y/%m/%d %H:%M", tz=pytz.UTC), "2014/01/02 01:04")
+
+    def test_date_range(self):
+        self.assertEqual(
+            [date(2015, 1, 29), date(2015, 1, 30), date(2015, 1, 31), date(2015, 2, 1)],
+            list(date_range(date(2015, 1, 29), date(2015, 2, 2))),
+        )
+        self.assertEqual([], list(date_range(date(2015, 1, 29), date(2015, 1, 29))))
 
 
 class CountriesTest(TembaTest):